requirements.txt
bs4
pyrogram==2.0.30
python-dotenv
requests
Tgcrypto
